Fix a bunch of implicit type conversion warnings that show up under Visual C++ and GCC when you add the -Wconversion option.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@920368 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/io/InputStream.cpp
@@ -179,18 +179,18 @@ int InputStream::doReadArrayBounded( unsigned char* buffer, std::size_t size,
             int c;
             try {
                 if( ( c = doReadByte() ) == -1 ) {
-                    return i == 0 ? -1 : i;
+                    return i == 0 ? -1 : (int)i;
                 }
             } catch( IOException& e ) {
                 if( i != 0 ) {
-                    return i;
+                    return (int)i;
                 }
                 throw e;
             }
             buffer[offset + i] = (unsigned char)c;
         }
 
-        return length;
+        return (int)length;
     }
     DECAF_CATCH_RETHROW( IOException )
     DECAF_CATCH_RETHROW( IndexOutOfBoundsException )
